ima: Move type defitions from ima_dm.py to types.py
